refactor(commands): commands.rs 拆为 commands/ 域模块（glob 重导出，lib.rs 零改）
把 4857 行的单文件 commands.rs 按单一职责拆成 commands/ 下 16 个域模块
（settings / credentials / providers / history / dictionary / dictation /
style_packs / permissions_cmds / qa / hotkeys / local_asr / foundry_asr /
sherpa_asr / marketplace / github_oauth / misc）。纯机械式、行为保持的剪切 +
可见性提升，未改动任何 #[tauri::command] 函数体。

- commands.rs → commands/mod.rs：内含 `mod <domain>;` + `pub use <domain>::*;`，
  glob 重导出每个域模块。`commands::<name>` 路径全部不变 → lib.rs 的
  generate_handler! 清单（346-492）与所有类型/helper 引用零改动。
  #[tauri::command] 生成的 __cmd__<name> 伴生项随 glob 一并重导出。
- 跨域共享项提到 mod.rs 并 pub(crate)：is_valid_session_id /
  is_valid_local_pack_id、open_path_in_file_manager（3 个 cfg 变体）、
  共享状态类型（MicrophoneMonitorState / TrayMicrophoneMenuState 等）、
  常用 crate 级 import 以 `pub(crate) use` 重导出供子模块 `use super::*` 取用。
- MARKETPLACE_BASE_URL 提为 pub(crate)（misc::check_network 引用）；
  hotkeys.rs 的 reject_* 整套 + sync_dictation_hotkey_legacy_fields 保留在
  hotkeys.rs 但提 pub(crate)（被 settings::persist_settings、qa::set_qa_hotkey 引用）；
  style_packs.rs 的 activate_builtin_style_mode 等保留并 pub(crate)（lib.rs 引用）。
- 测试：保留单个 tests mod 在 mod.rs，把 `use super::{...}` 改为 per-module
  路径，被测 helper 提为 pub(crate)。不改任何测试逻辑。
- sherpa_onnx_asr_* 整组 #[cfg(windows)]，非 Windows 平台 glob 重导出无人引用，
  对该 `pub use sherpa_asr::*` 加 #[allow(unused_imports)]。

验证：cargo check 0 error；cargo test --lib 417 passed（基线不回归）；
npm run build 通过；capsule Spaces / hotkey-injection 契约测试通过。
命令面 127 个 #[tauri::command] 与原文件逐一对齐，无新增/缺失/重复。
