+++ openless-all/app/src-tauri/src/commands/dictation.rs
@@ -1,0 +1,49 @@
+use super::*;
+
+#[tauri::command]
+pub async fn start_dictation(coord: CoordinatorState<'_>) -> Result<(), String> {
+    coord.start_dictation().await
+}
+
+#[tauri::command]
+pub async fn stop_dictation(coord: CoordinatorState<'_>) -> Result<(), String> {
+    coord.stop_dictation().await
+}
+
+#[tauri::command]
+pub fn cancel_dictation(coord: CoordinatorState<'_>) {
+    coord.cancel_dictation();
+}
+
+#[tauri::command]
+pub async fn handle_window_hotkey_event(
+    coord: CoordinatorState<'_>,
+    event_type: String,
+    key: String,
+    code: String,
+    repeat: bool,
+) -> Result<(), String> {
+    coord
+        .handle_window_hotkey_event(event_type, key, code, repeat)
+        .await
+}
+
+#[cfg(debug_assertions)]
+#[tauri::command]
+pub async fn inject_hotkey_click_for_dev(coord: CoordinatorState<'_>) -> Result<(), String> {
+    coord.inject_hotkey_click_for_dev().await
+}
+
+#[tauri::command]
+pub async fn repolish(
+    coord: CoordinatorState<'_>,
+    raw_text: String,
+    mode: PolishMode,
+) -> Result<String, String> {
+    log::info!(
+        "[style-pack] command repolish requested legacy_mode={:?} raw_chars={}",
+        mode,
+        raw_text.chars().count()
+    );
+    coord.repolish(raw_text, mode).await
+}
