+++ openless-all/app/src-tauri/src/commands/foundry_asr.rs
@@ -1,0 +1,190 @@
+use super::*;
+
+pub(crate) fn active_foundry_model_from_prefs(prefs: &UserPreferences) -> String {
+    if model_alias_is_known(&prefs.foundry_local_asr_model) {
+        prefs.foundry_local_asr_model.clone()
+    } else {
+        DEFAULT_MODEL_ALIAS.to_string()
+    }
+}
+
+pub(crate) fn validate_foundry_model_alias(model_alias: &str) -> Result<(), String> {
+    if model_alias_is_known(model_alias) {
+        Ok(())
+    } else {
+        Err(format!(
+            "unknown Foundry Whisper model alias: {model_alias}"
+        ))
+    }
+}
+
+pub(crate) fn normalize_foundry_language_hint(language_hint: &str) -> Result<String, String> {
+    let normalized = language_hint.trim().to_string();
+    if normalized.is_empty()
+        || (normalized.len() == 2 && normalized.bytes().all(|b| b.is_ascii_lowercase()))
+    {
+        Ok(normalized)
+    } else {
+        Err("language hint must be empty or ISO 639-1 lowercase code".to_string())
+    }
+}
+
+fn normalize_foundry_runtime_source(source: &str) -> String {
+    crate::asr::local::foundry_native::normalize_runtime_source_str(source)
+}
+
+#[tauri::command]
+pub async fn foundry_local_asr_status(
+    coord: CoordinatorState<'_>,
+    runtime: State<'_, Arc<FoundryLocalRuntime>>,
+) -> Result<FoundryRuntimeStatus, String> {
+    let prefs = coord.prefs().get();
+    let active_model = active_foundry_model_from_prefs(&prefs);
+    Ok(runtime
+        .status_snapshot(&active_model, &prefs.foundry_local_runtime_source)
+        .await)
+}
+
+#[tauri::command]
+pub async fn foundry_local_asr_catalog(
+    runtime: State<'_, Arc<FoundryLocalRuntime>>,
+) -> Result<Vec<FoundryCatalogModel>, String> {
+    runtime
+        .catalog_snapshot()
+        .await
+        .map_err(|e| format!("{e:#}"))
+}
+
+#[tauri::command]
+pub fn foundry_local_asr_set_model(
+    coord: CoordinatorState<'_>,
+    model_alias: String,
+) -> Result<(), String> {
+    validate_foundry_model_alias(&model_alias)?;
+    let mut prefs = coord.prefs().get();
+    if prefs.foundry_local_asr_model == model_alias {
+        return Ok(());
+    }
+    prefs.foundry_local_asr_model = model_alias;
+    coord.prefs().set(prefs).map_err(|e| e.to_string())
+}
+
+#[tauri::command]
+pub fn foundry_local_asr_set_language_hint(
+    coord: CoordinatorState<'_>,
+    language_hint: String,
+) -> Result<(), String> {
+    let normalized = normalize_foundry_language_hint(&language_hint)?;
+    let mut prefs = coord.prefs().get();
+    if prefs.foundry_local_asr_language_hint == normalized {
+        return Ok(());
+    }
+    prefs.foundry_local_asr_language_hint = normalized;
+    coord.prefs().set(prefs).map_err(|e| e.to_string())
+}
+
+#[tauri::command]
+pub fn foundry_local_asr_set_runtime_source(
+    coord: CoordinatorState<'_>,
+    source: String,
+) -> Result<(), String> {
+    let mut prefs = coord.prefs().get();
+    let normalized = normalize_foundry_runtime_source(&source);
+    if prefs.foundry_local_runtime_source == normalized {
+        return Ok(());
+    }
+    prefs.foundry_local_runtime_source = normalized;
+    coord.prefs().set(prefs).map_err(|e| e.to_string())
+}
+
+#[tauri::command]
+pub async fn foundry_local_asr_prepare(
+    app: AppHandle,
+    coord: CoordinatorState<'_>,
+    runtime: State<'_, Arc<FoundryLocalRuntime>>,
+    model_alias: String,
+) -> Result<String, String> {
+    validate_foundry_model_alias(&model_alias)?;
+    let prefs = coord.prefs().get();
+    let runtime_source = prefs.foundry_local_runtime_source.clone();
+    let progress_app = app.clone();
+    let result = runtime
+        .ensure_loaded_with_progress(&model_alias, &runtime_source, move |payload| {
+            emit_foundry_prepare_progress(&progress_app, payload);
+        })
+        .await;
+    match result {
+        Ok(model_id) => Ok(model_id),
+        Err(error) => {
+            let message = format!("{error:#}");
+            emit_foundry_prepare_progress(
+                &app,
+                FoundryPrepareProgressPayload::failed(
+                    model_alias,
+                    "Foundry Local Whisper prepare failed",
+                    message.clone(),
+                ),
+            );
+            Err(message)
+        }
+    }
+}
+
+#[tauri::command]
+pub fn foundry_local_asr_cancel_prepare(
+    runtime: State<'_, Arc<FoundryLocalRuntime>>,
+) -> Result<(), String> {
+    runtime.request_cancel_prepare();
+    Ok(())
+}
+
+#[tauri::command]
+pub async fn foundry_local_asr_release(
+    runtime: State<'_, Arc<FoundryLocalRuntime>>,
+) -> Result<(), String> {
+    runtime.release_now().await.map_err(|e| format!("{e:#}"))
+}
+
+#[tauri::command]
+pub async fn foundry_local_asr_model_dir(
+    runtime: State<'_, Arc<FoundryLocalRuntime>>,
+    model_alias: String,
+) -> Result<String, String> {
+    validate_foundry_model_alias(&model_alias)?;
+    runtime
+        .model_dir_for_alias(&model_alias)
+        .await
+        .map(|path| path.display().to_string())
+        .map_err(|e| format!("{e:#}"))
+}
+
+#[tauri::command]
+pub async fn foundry_local_asr_delete_model(
+    runtime: State<'_, Arc<FoundryLocalRuntime>>,
+    model_alias: String,
+) -> Result<(), String> {
+    validate_foundry_model_alias(&model_alias)?;
+    runtime
+        .delete_model(&model_alias)
+        .await
+        .map_err(|e| format!("{e:#}"))
+}
+
+#[tauri::command]
+pub async fn foundry_local_asr_reveal_model_dir(
+    runtime: State<'_, Arc<FoundryLocalRuntime>>,
+    model_alias: String,
+) -> Result<(), String> {
+    validate_foundry_model_alias(&model_alias)?;
+    let dir = runtime
+        .model_dir_for_alias(&model_alias)
+        .await
+        .map_err(|e| format!("{e:#}"))?;
+    open_path_in_file_manager(&dir)
+}
+
+fn emit_foundry_prepare_progress(app: &AppHandle, payload: FoundryPrepareProgressPayload) {
+    if let Err(error) = app.emit("foundry-local-asr-prepare-progress", payload) {
+        log::warn!("[foundry-asr] emit prepare progress failed: {error}");
+    }
+}
