+++ openless-all/app/src-tauri/src/commands/history.rs
@@ -1,0 +1,50 @@
+use super::*;
+
+#[tauri::command]
+pub fn list_history(coord: CoordinatorState<'_>) -> Result<Vec<DictationSession>, String> {
+    coord.history().list().map_err(|e| e.to_string())
+}
+
+#[tauri::command]
+pub fn delete_history_entry(coord: CoordinatorState<'_>, id: String) -> Result<(), String> {
+    coord.history().delete(&id).map_err(|e| e.to_string())
+}
+
+#[tauri::command]
+pub fn clear_history(coord: CoordinatorState<'_>) -> Result<(), String> {
+    coord.history().clear().map_err(|e| e.to_string())
+}
+
+/// 读取某次会话的原始麦克风 wav 字节流。仅当用户开过
+/// `prefs.record_audio_for_debug` 并且这条 session 是开关打开后录的，才会有文件。
+/// 文件名规约：`<data_dir>/recordings/<session_id>.wav`，与 DictationSession.id 同名。
+///
+/// 路径校验：session_id **必须**严格匹配 UUID-v4 字面（36 字符 = 8-4-4-4-12 + 4 个 `-`，
+/// 内容仅 ASCII 十六进制 + `-`）。白名单胜过黑名单——绝对路径前缀、Windows ADS、
+/// 百分号编码、NUL 字节都不在合法字符集里，挡掉所有 Path::join 越界的可能。
+/// session_id 在仓库内由 `Uuid::new_v4()` 生成 (`dictation.rs:1531`)，前端只会回传
+/// 自己列出的合法 id，但 IPC = boundary，按 boundary 规则严格校验。
+///
+/// async fs：单条 5 分钟 wav 约 9.6MB，同步 `std::fs::read` 会阻塞 Tauri IPC 主循环。
+/// 改 `tokio::fs::read` 后让出线程给其它 IPC。
+#[tauri::command]
+pub async fn read_audio_recording(session_id: String) -> Result<Vec<u8>, String> {
+    if !is_valid_session_id(&session_id) {
+        return Err("invalid session id".into());
+    }
+    let path =
+        crate::persistence::recording_path_for_session(&session_id).map_err(|e| e.to_string())?;
+    if !path.exists() {
+        return Err("recording not found".into());
+    }
+    // TOCTOU 兜底：exists() 通过到 read 之间文件可能被 prune（条数 cap / retention
+    // 清理 / 用户手动删）。把 NotFound 标准化成跟 exists() 失败同样的错误字符串，
+    // 前端单条 'recording not found' catch 就能稳定隐藏按钮，不依赖本地化 OS 错误。
+    tokio::fs::read(&path).await.map_err(|e| {
+        if e.kind() == std::io::ErrorKind::NotFound {
+            "recording not found".into()
+        } else {
+            format!("read wav failed: {e}")
+        }
+    })
+}
