+++ openless-all/app/src-tauri/src/commands/permissions_cmds.rs
@@ -1,0 +1,100 @@
+use super::*;
+
+#[tauri::command]
+pub fn check_accessibility_permission() -> PermissionStatus {
+    permissions::check_accessibility()
+}
+
+#[tauri::command]
+pub fn request_accessibility_permission() -> PermissionStatus {
+    permissions::request_accessibility()
+}
+
+#[tauri::command]
+pub fn check_microphone_permission() -> PermissionStatus {
+    permissions::check_microphone()
+}
+
+#[tauri::command]
+pub fn request_microphone_permission(app: AppHandle) -> PermissionStatus {
+    crate::request_microphone_from_foreground(&app)
+}
+
+/// 跳到 macOS 系统设置的指定隐私面板。pane: "accessibility" | "microphone".
+#[tauri::command]
+pub fn open_system_settings(pane: String) -> Result<(), String> {
+    #[cfg(target_os = "macos")]
+    {
+        let url = match pane.as_str() {
+            "accessibility" => {
+                "x-apple.systempreferences:com.apple.preference.security?Privacy_Accessibility"
+            }
+            "microphone" => {
+                "x-apple.systempreferences:com.apple.preference.security?Privacy_Microphone"
+            }
+            _ => "x-apple.systempreferences:com.apple.preference.security?Privacy",
+        };
+        std::process::Command::new("open")
+            .arg(url)
+            .spawn()
+            .map(|_| ())
+            .map_err(|e| e.to_string())
+    }
+    #[cfg(target_os = "windows")]
+    {
+        use windows::core::PCWSTR;
+        use windows::Win32::UI::Shell::ShellExecuteW;
+        use windows::Win32::UI::WindowsAndMessaging::SW_SHOWNORMAL;
+
+        fn wide_null(value: &str) -> Vec<u16> {
+            value.encode_utf16().chain(std::iter::once(0)).collect()
+        }
+
+        let uri = match pane.as_str() {
+            "microphone" => "ms-settings:privacy-microphone",
+            "sound" => "ms-settings:sound",
+            "accessibility" => "ms-settings:easeofaccess",
+            _ => "ms-settings:",
+        };
+
+        let operation = wide_null("open");
+        let target = wide_null(uri);
+        let result = unsafe {
+            ShellExecuteW(
+                None,
+                PCWSTR(operation.as_ptr()),
+                PCWSTR(target.as_ptr()),
+                PCWSTR::null(),
+                PCWSTR::null(),
+                SW_SHOWNORMAL,
+            )
+        };
+
+        if result.0 as isize <= 32 {
+            Err(format!("ShellExecuteW failed: {}", result.0 as isize))
+        } else {
+            Ok(())
+        }
+    }
+    #[cfg(all(not(target_os = "macos"), not(target_os = "windows")))]
+    {
+        let _ = pane;
+        Err("open_system_settings is only supported on macOS and Windows".to_string())
+    }
+}
+
+/// 触发 macOS 系统弹"是否允许 OpenLess 访问麦克风"对话框。
+/// 与 Swift `MicrophonePermission.request()` 同语义：只信系统权限回调，
+/// 不用 cpal stream 成功与否伪造授权状态。
+#[tauri::command]
+pub fn trigger_microphone_prompt(app: AppHandle) -> Result<(), String> {
+    let status = crate::request_microphone_from_foreground(&app);
+    if matches!(
+        status,
+        PermissionStatus::Granted | PermissionStatus::NotApplicable
+    ) {
+        Ok(())
+    } else {
+        Err(format!("microphone permission is {status:?}"))
+    }
+}
