+++ openless-all/app/src-tauri/src/commands/qa.rs
@@ -1,0 +1,67 @@
+use super::*;
+
+#[tauri::command]
+pub fn get_qa_hotkey_label(coord: CoordinatorState<'_>) -> String {
+    coord.qa_hotkey_label()
+}
+
+/// 设置 QA 快捷键并热更新 monitor。
+/// 传入 `None` 形式的字段不在这里支持——前端用 `binding == null` 时调下面的
+/// "disable" 写法（写 prefs.qa_hotkey = None）即可。
+#[tauri::command]
+pub fn set_qa_hotkey(
+    coord: CoordinatorState<'_>,
+    binding: Option<ShortcutBinding>,
+) -> Result<(), String> {
+    if let Some(binding) = binding.as_ref() {
+        crate::shortcut_binding::validate_binding(binding).map_err(|e| e.to_string())?;
+        if binding.modifiers.is_empty() && binding.primary.eq_ignore_ascii_case("shift") {
+            return Err("Shift 单键目前只能用于翻译快捷键".into());
+        }
+    }
+    let mut prefs = coord.prefs().get();
+    if let Some(binding) = binding.as_ref() {
+        reject_dictation_qa_hotkey_overlap(&prefs.dictation_hotkey, binding)?;
+        reject_qa_translation_hotkey_overlap(binding, &prefs.translation_hotkey)?;
+        if let Some(switch_style) = prefs.switch_style_hotkey.as_ref() {
+            reject_qa_switch_style_hotkey_overlap(binding, switch_style)?;
+        }
+        if let Some(open_app) = prefs.open_app_hotkey.as_ref() {
+            reject_qa_open_app_hotkey_overlap(binding, open_app)?;
+        }
+    }
+    prefs.qa_hotkey = binding;
+    coord.prefs().set(prefs).map_err(|e| e.to_string())?;
+    coord.update_qa_hotkey_binding();
+    Ok(())
+}
+
+/// 用户点 ✕ / 按 Esc 关 QA 浮窗。
+#[tauri::command]
+pub fn qa_window_dismiss(coord: CoordinatorState<'_>) {
+    coord.qa_window_dismiss();
+}
+
+/// 用户点 📌 / 取消 📌。pinned=true 时浮窗不会自动隐藏。
+#[tauri::command]
+pub fn qa_window_pin(coord: CoordinatorState<'_>, pinned: bool) {
+    coord.qa_window_pin(pinned);
+}
+
+/// 用户点 ✕ / 按 Esc 关 Less Computer 浮窗。
+#[tauri::command]
+pub fn less_computer_window_dismiss(coord: CoordinatorState<'_>) {
+    coord.less_computer_window_dismiss();
+}
+
+/// 前端按内容测高后回传高度，后端 clamp + bottom-anchored 重新摆放浮窗。
+#[tauri::command]
+pub fn less_computer_window_resize(coord: CoordinatorState<'_>, height: f64) {
+    coord.less_computer_window_resize(height);
+}
+
+/// 内联审批卡的 Approve / Deny 回执。token 关联到等待中的拦截动作。
+#[tauri::command]
+pub fn less_computer_approve(coord: CoordinatorState<'_>, token: String, approved: bool) {
+    coord.less_computer_approve(&token, approved);
+}
